+++ SoftWare/STM32G431CBUx/motorlib/bsp/stm32_as5047.h
@@ -1,6 +1,6 @@
 #ifndef STM32_AS5047_H
 #define STM32_AS5047_H
 #include "stdint.h"
-uint16_t as5047_read_raw(void);
+uint32_t as5047_read_raw(void);
 
 #endif
